+++ src/server/server_main_tab.R
@@ -471,6 +471,7 @@ updateLevels <- reactive({
   input$station
   updateVariables()
   reloadInfoFromCache()
+  activePlotType()
   if(length(availableLevels())==0) invalidateLater(500)
 }) %>% throttle(500)
 observeEvent({
@@ -487,12 +488,16 @@ observeEvent({
   )
 
   # Toggle the choice between all levels or standard-only
-  hasStandardLevels <- length(levels$obsmon) > 0
-  allLevelsAreStandard <- all(levels$all %in% levels$obsmon)
-  showStandardLevelsToggle <- hasStandardLevels && !allLevelsAreStandard
-  shinyjs::toggle("standardLevelsSwitch", condition=showStandardLevelsToggle)
+  queryFromUsageDbTable <- activePlotType()$queriedDbTable == "usage"
+  usageHasStandardLevels <- length(intersect(levels$obsmon, levels$usage)) > 0
+  allUsageLevelsAreStandard <- all(levels$usage %in% levels$obsmon)
   shinyjs::toggle(
-    "groupLevelsIntoStandardSwitch", condition=!allLevelsAreStandard
+    "standardLevelsSwitch",
+    condition=queryFromUsageDbTable && usageHasStandardLevels && !allUsageLevelsAreStandard
+  )
+  shinyjs::toggle(
+    "groupLevelsIntoStandardSwitch",
+    condition=queryFromUsageDbTable && !allUsageLevelsAreStandard
   )
 
   availableLevels(levels)
@@ -503,9 +508,22 @@ observeEvent({
   availableLevels()
   input$standardLevelsSwitch
   levelsUnitsChanged()
+  activePlotType()
 }, {
-    if(isTRUE(input$standardLevelsSwitch)) choices <- availableLevels()$obsmon
-    else choices <- availableLevels()$all
+    queryFromUsageDbTable <- activePlotType()$queriedDbTable == "usage"
+    if(queryFromUsageDbTable) {
+      availableStandardLevels <- intersect(
+        availableLevels()$obsmon,
+        availableLevels()$usage
+      )
+      if(isTRUE(input$standardLevelsSwitch)) {
+        choices <- availableStandardLevels
+      } else {
+        choices <- unique(c(availableStandardLevels, availableLevels()$usage))
+      }
+    } else {
+      choices <- availableLevels()$obsmon
+    }
 
     # Present level choices in the units picked by the user, but make sure
     # to pass it to the query with the expected (default) units